// book-socials-ui/src/app/services/models/borrowed-book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BorrowedBookResponse {
  authorName?: string;
  borrowedBy?: string;
  borrowedByUserUuid?: string;
  isbn?: string;
  rate?: number;
  returnApproved?: boolean;
  returned?: boolean;
  title?: string;
  uuid?: string;
}