// book-socials-ui/src/app/services/models/page-response-borrowed-book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BorrowedBookResponse } from '../models/borrowed-book-response';
export interface PageResponseBorrowedBookResponse {
  first?: boolean;
  hydraMember?: Array<BorrowedBookResponse>;
  hydraTotal?: number;
  last?: boolean;
  page?: number;
  size?: number;
  totalPages?: number;
}